requirements.txt

mysql-connector-python
pandas
nltk
spacy
pytest
jupyter